+++ WindowsApp.h
@@ -34,7 +34,7 @@ private:
 		//ウインドウクラスをOSに登録する
 		RegisterClassEx(&w);
 		//ウインドウサイズ{X,Y,横幅,縦幅}
-		RECT wrc = { 0,0,window_width,window_height };
+		RECT wrc = { 0,0,(LONG)window_width,(LONG)window_height };
 		//自動でサイズ補正
 		AdjustWindowRect(&wrc, WS_OVERLAPPEDWINDOW, false);
